+++ src/constants/appVersion.ts
@@ -1,1 +1,1 @@
-export const APP_VERSION = "13.141.8";
+export const APP_VERSION = "13.141.9";